api/requirement.txt
alembic==1.7.1
amqp==5.0.6
autopep8==1.5.7
billiard==3.6.4.0
blinker==1.4
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.0.1
google-api-python-client==2.19.1
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.1
httplib2==0.19.1
idna==3.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
oauthlib==3.1.1
Pillow==8.3.2
prompt-toolkit==3.0.20
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-ffmpeg-video-streaming==0.1.14
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
stripe==2.60.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
gunicorn
